refactor(NONE): 통계 쿼리 SolveHistoryStatisticsRepository로 분리
